Moved the navbar components into a shared folder
