+++ frontend/src/lib/components/NavBarComponents/index.ts
@@ -1,2 +1,3 @@
 export type { NavBarButtonProps } from "./navBarButton";
 export { NavBarButton } from "./navBarButton";
+export { NavBarDivider } from "./navBarDivider";
