+++ frontend/src/lib/components/NavBarComponents/navBarDivider.tsx
@@ -1,0 +1,5 @@
+import type React from "react";
+
+export function NavBarDivider(): React.ReactNode {
+    return <div className="bg-slate-200 h-px w-full mt-4 mb-4" />;
+}
